+++ src/components/mypage/EditMyPage/LogoutModal.tsx
@@ -1,4 +1,4 @@
-import ReusableModal from '../common/Modal/ReusableModal';
+import ReusableModal from '../../common/Modal/ReusableModal';
 
 interface LogoutModalProps {
   isOpen: boolean;
