+++ src/components/mypage/ResetPassword/TemporaryPasswordModal.tsx
