+++ src/components/phototalk/Card/PhotoTalkAction.tsx
