+++ src/components/phototalk/Card/PhotoTalkMessage.tsx
