+++ src/components/phototalk/EmptyState/PhotoTalkEmptyState.tsx
