+++ src/components/phototalk/Gallery/PhotoTalkGallery.tsx
@@ -1,6 +1,6 @@
-import PhotoTalkGalleryAdmin from '@/components/common/PhotoTalk/Gallery/PhotoTalkGalleryAdmin';
-import PhotoTalkGalleryGuest from '@/components/common/PhotoTalk/Gallery/PhotoTalkGalleryGuest';
-import PhotoTalkGalleryPreview from '@/components/common/PhotoTalk/Gallery/PhotoTalkGalleryPreview';
+import PhotoTalkGalleryAdmin from '@/components/phototalk/Gallery/PhotoTalkGalleryAdmin';
+import PhotoTalkGalleryGuest from '@/components/phototalk/Gallery/PhotoTalkGalleryGuest';
+import PhotoTalkGalleryPreview from '@/components/phototalk/Gallery/PhotoTalkGalleryPreview';
 import { exampleImages } from '@/constants/phototalkData';
 import usePhotoTalkStore from '@/store/usePhotoTalkStore';
 import { USER_MODE, UserMode } from '@/types/users';
