+++ src/components/phototalk/Gallery/PhotoTalkGalleryPreview.tsx
@@ -1,5 +1,5 @@
-import PhotoTalkGalleryGrid from '@/components/common/PhotoTalk/Gallery/PhotoTalkGalleryGrid';
-import PhotoTalkGalleryModal from '@/components/common/PhotoTalk/Modal/GalleryModal/PhotoTalkGalleryModal';
+import PhotoTalkGalleryGrid from '@/components/phototalk/Gallery/PhotoTalkGalleryGrid';
+import PhotoTalkGalleryModal from '@/components/phototalk/Modal/GalleryModal/PhotoTalkGalleryModal';
 import { phototalkData } from '@/constants/phototalkData';
 import { UserMode } from '@/types/users';
 import { useState } from 'react';
