+++ src/components/phototalk/Modal/GalleryModal/PhotoTalkGalleryModalFooter.tsx
